Moved date-time related utils to the proper folder
