+++ src/utils/dates/DateInput.scss
@@ -1,5 +1,5 @@
-@import './mixins/vertical-align';
-@import './base';
+@import '../mixins/vertical-align';
+@import '../base';
 
 .date-input-container {
   position: relative;
