+++ src/utils/dates/DateTimeInput.tsx
@@ -1,6 +1,6 @@
 import { ReactDatePickerProps } from 'react-datepicker';
 import { FC } from 'react';
-import { DateInput } from '../DateInput';
+import { DateInput } from './DateInput';
 import { STANDARD_DATE_AND_TIME_FORMAT } from '../helpers/date';
 
 export type DateTimeInputProps = Omit<ReactDatePickerProps, 'showTimeSelect' | 'dateFormat' | 'timeIntervals'>;
